refactor: split ReportingMutationTests into one file per nested suite
